% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MSPQ_analysisV1_ONLY_RANK.R
\name{MSPQ_ranks}
\alias{MSPQ_ranks}
\title{MSPQ_ranks}
\usage{
MSPQ_ranks(
  out,
  perIter = 100,
  PerSeed = 123,
  spats = FALSE,
  row = NULL,
  column = NULL,
  pl.date = FALSE
)
}
\arguments{
\item{out}{output}

\item{perIter}{iterations}

\item{PerSeed}{seed}

\item{spats}{TRUE or FALSE if spatial analysis is needed}

\item{row}{row information}

\item{column}{column information}

\item{pl.date}{date}
}
\value{
list
}
\description{
MSPQ_ranks
}
\examples{
\dontrun{
MSPQ_ranks()
}
}
