% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MSPQ_tidyV2.R
\name{MSPQ_tidy}
\alias{MSPQ_tidy}
\title{MSPQ_tidy}
\usage{
MSPQ_tidy(df, genotype, time.diff, data_name = NULL, plotIm = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{genotype}{string}

\item{time.diff}{logical}

\item{data_name}{string}

\item{plotIm}{logical}
}
\value{
list
}
\description{
MSPQ_tidy
}
\examples{
\dontrun{
MSPQ_tidy()
}
}
