% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_yield_calibration.R
\name{target_trait_comparison}
\alias{target_trait_comparison}
\title{Comparisons between raknings and target trait of interest.}
\usage{
target_trait_comparison(ranks, target.trait.file, target.trait.name, metadata = NULL)
}
\arguments{
\item{ranks}{The list obtained by MSPQ_ranks.}

\item{target.trait.file}{data.frame in narrow format with the yield component or any other trait to contrast against MultispeQ ranking. An observation per experimental unit is required as well as the row & column coordinates if spats = TRUE in MSPQ_ranks.}

\item{target.trait.name}{Character. name of the yield variable on yield.file to contrast against MultispeQ ranking.}

\item{metadata}{data.frame in narrow format with extra information about the genotypes i.e. gene pool, known biotic/abiotic traits, etc. Optional process. NULL by default.}
}
\value{
A list with the confusion matrices and summary of predictions.
}
\description{
target_trait_comparison contrasts the previous genotype-trait classification of MultispeQ measurements against final yield or any other trait of interest of the evaluated genotypes.
}
\examples{
\dontrun{
# yield_data: any table containing the same evaluated genotypes and the yield or any target to contrast against MSPQ rankings

target_trait_comparison(ranks = ranks,
                        target.trait.file = yield_data,
                        target.trait.name = kg_ha,
                        metadata = NULL)
}
}
