% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_yield_calibration.R
\name{yield_calibration}
\alias{yield_calibration}
\title{Comparisons between raknings and yield or trait of interest.}
\usage{
yield_calibration(ranks, yield.file, yield.name, metadata = NULL)
}
\arguments{
\item{ranks}{The list obtained by MSPQ_ranks.}

\item{yield.file}{data.frame in narrow format with the yield component or any other trait to contrast against MultispeQ ranking. An observation per experimental unit is required as well as the row & column coordinates if spats = TRUE in MSPQ_ranks.}

\item{yield.name}{Character. name of the yield variable on yield.file to contrast against MultispeQ ranking.}

\item{metadata}{data.frame in narrow format with extra information about the genotypes i.e. gene pool, known biotic/abiotic traits, etc. Optional process. NULL by default.}
}
\value{
A list with the confusion matrices and summary of predictions.
}
\description{
yield_calibration contrasts the previous genotype-trait classification of MultispeQ measurements against final yield or any other yield component of the evaluated genotypes.
}
\examples{
\dontrun{
# yield_data: any table containing the same evaluated genotypes and the yield or any trait of interest to contrast against rankings

yield_calibration(ranks = ranks,
                  yield.file = yield_data,
                  yield.name = kg_ha,
                  metadata = NULL)
}
}
