% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_MSPQ_tidyV2.R
\name{MSPQ_tidy}
\alias{MSPQ_tidy}
\title{Tidying and preparing the MultispeQ raw dataset.}
\usage{
MSPQ_tidy(df, genotype, time.diff, data_name = NULL, plotIm = FALSE)
}
\arguments{
\item{df}{data.frame The raw data set obtained from the PhotosynQ network.}

\item{genotype}{string with the name of the genotype column on df.}

\item{time.diff}{logical. Were the measurements repeated twice in a day? Morning & Afternoon. TRUE if so, FALSE otherwise}

\item{data_name}{string with the name of the experiment.}

\item{plotIm}{logical. To plot the imputed variables. FALSE by default.}
}
\value{
A list with 8 objects if plotIm = FALSE. 9 elements otherwise.
}
\description{
MSPQ_tidy prepares the raw data set for the further analysis and genotype-trait classification. It also provides a summary with the outlier detection and variables imputation in case of so.
}
\examples{
\dontrun{
#MSPQ_csv_data: any csv dataset obtained from the PhotosynQ Network.

tidy_data <- MSPQ_tidy(df = MSPQ_csv_data,
                       genotype = "Genotype",
                       time.diff = FALSE,
                       data_name = "My_Project",
                       plotIm = TRUE)
}
}
