% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_yield_calibration.R
\name{yield_calibration}
\alias{yield_calibration}
\title{yield_calibration}
\usage{
yield_calibration(ranks, yield.file, yield.name, metadata = NULL)
}
\arguments{
\item{ranks}{ranks}

\item{yield.file}{file}

\item{yield.name}{name}

\item{metadata}{metadata}
}
\value{
list
}
\description{
yield_calibration
}
\examples{
\dontrun{
yield_calibration()
}
}
